Bring HfpProtocol back
